// ui/src/app/common/iam-client-react/action-types.ts
export enum ActionTypes {
    REQUEST_LOGIN = 'REQUEST_LOGIN',
    LOGIN_SUCCESS = 'LOGIN_SUCCESS',
    LOGIN_ERROR = 'LOGIN_ERROR',
    LOGOUT = 'LOGOUT',
    ASSIGN_UNIT = 'ASSIGN_UNIT',
    FORGOT_PASSWORD = 'FORGOT_PASSWORD',
    REQUEST_OTP = 'REQUEST_OTP',
    OTP_SUCCESS = 'OTP_SUCCESS',
    OTP_ERROR = 'OTP_ERROR',
    REQUEST_RESET_PWD = 'REQUEST_RESET_PWD',
    RESET_PWD_SUCCESS = 'RESET_PWD_SUCCESS',
    RESET_PWD_ERROR = 'OTP_ERROR',
}